Applied patch from https://bugs.eclipse.org/bugs/attachment.cgi?id=108454

+++ org.eclipse.buckminster.download/src/org/eclipse/buckminster/download/internal/FileReader.java
@@ -78,7 +78,7 @@ public class FileReader extends FileTransferJob implements IFileTransferListener
 		setSystem(true);
 		setUser(false);
 		m_connectionRetryCount = BuckminsterPreferences.getConnectionRetryCount();
-		m_connectionRetryDelay = (long)BuckminsterPreferences.getConnectionRetryDelay() * 1000L;
+		m_connectionRetryDelay = BuckminsterPreferences.getConnectionRetryDelay() * 1000L;
 	}
 
 	public IFileInfo getLastFileInfo()
